moved main module back out of file
