+++ sublease-app/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {Routes, RouterModule} from '@angular/router';
-import { AuthenticateComponent } from '../Components/Forms/authenticate/authenticate.component';
+import { AuthenticateComponent } from './Components/Forms/authenticate/authenticate.component';
 
 
 const routes: Routes = [
