+++ sublease-app/src/app/app.module.ts
@@ -1,10 +1,10 @@
-import { MaterialModule } from '../Materials/MaterialsModule';
+import { MaterialModule } from './Materials/MaterialsModule';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import {HttpClientModule} from '@angular/common/http';
 import { AppComponent } from './app.component';
-import { AuthenticateComponent } from '../Components/Forms/authenticate/authenticate.component';
+import { AuthenticateComponent } from './Components/Forms/authenticate/authenticate.component';
 import { AppRoutingModule } from './app-routing.module';
 
 @NgModule({
